Network Observer Added (#1)
* Create NetworkObserver.kt

This component is a simple working implementation of a NetworkCallback that manages events from the system and determines whether there is a network connection that satisfies NetworkRequest. Override buildNetworkRequest to narrow the results received from the system to networks with specific properties. RegisterWith and UnregisterWith take Context as an input for convenience.

* Update ContextHelp.kt

Removed the getConnectivityMgr convenience method after updating the NetworkCallback class.

* Update README.md

Added Context package information and information on NetworkObserver usage.